cr: move RuntimeEnvironment to constants
