+++ src/setNextRuntimeConfig.ts
@@ -1,5 +1,5 @@
 import { setConfig } from 'next/dist/next-server/lib/runtime-config';
-import { RuntimeEnvironment } from './commonTypes';
+import { RuntimeEnvironment } from './constants';
 import { getNextConfig } from './nextConfig';
 
 export default function setNextRuntimeConfig({
